refactor: restructure PlanFormTable component and update import paths
