+++ src/components/registration/plan-table/styled.ts
